course

+++ src/codeplus_java/BOJ1000_a.java
@@ -5,9 +5,9 @@ import java.util.*;
 public class BOJ1000_a {
 	public static void main(String args[]) {
 		Scanner sc = new Scanner(System.in);
-		int sum = 0;
-		while (sc.hasNextInt()) {
-			sum += sc.nextInt();
+		float sum = 0;
+		while (sc.hasNextFloat()) {
+			sum += sc.nextFloat();
 		}
 		System.out.println(sum);
 	}
